added xyz files

+++ src/neighbors.cpp
@@ -131,7 +131,7 @@ NeighborList::update(const Atoms &atoms) {
        return neighborhood;
    }();
 
-   for (int i{0}; i < atoms.nb_atoms(); ++i) {
+   for (int i{0}; i < (int)atoms.nb_atoms(); ++i) {
        seed_(i) = n;
 
        Eigen::Array3i cell_coord{
